# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agente-IA — contrato canónico",
    "version": "0.4.0",
    "description": "Fase 4: API operativa interna IMPLEMENTED; sin despliegue público ni OpenAI ni Meta outbound."
  },
  "paths": {
    "/v1/me": {
      "get": {
        "operationId": "get_v1_me",
        "summary": "Identidad y negocios autorizados",
        "description": "IMPLEMENTED en fase 4; sin despliegue público. Autorización y validación en docs/api/contracts.md.",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "supabaseBearer": []
          }
        ],
        "x-roles": [
          "authenticated"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Resultado propuesto; ver reglas del dominio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "x-implementation-status": "IMPLEMENTED"
      }
    },
    "/v1/businesses/{business_id}": {
      "get": {
        "operationId": "get_v1_businesses_business_id",
        "summary": "Consultar negocio",
        "description": "IMPLEMENTED en fase 4; sin despliegue público. Autorización y validación en docs/api/contracts.md.",
        "tags": [
          "business"
        ],
        "security": [
          {
            "supabaseBearer": []
          }
        ],
        "x-roles": [
          "owner",
          "manager",
          "operator"
        ],
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado propuesto; ver reglas del dominio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "x-implementation-status": "IMPLEMENTED"
      },
      "patch": {
        "operationId": "patch_v1_businesses_business_id",
        "summary": "Actualizar negocio",
        "description": "IMPLEMENTED en fase 4; sin despliegue público. Autorización y validación en docs/api/contracts.md.",
        "tags": [
          "business"
        ],
        "security": [
          {
            "supabaseBearer": []
          }
        ],
        "x-roles": [
          "owner",
          "manager"
        ],
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado propuesto; ver reglas del dominio.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            },
            "headers": {
              "Idempotency-Replayed": {
                "description": "true si se devolvió resultado persistido de mismo actor/comando.",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "x-implementation-status": "IMPLEMENTED",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessUpdate"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "supabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Error400": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Error401": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Error403": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Error404": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Error409": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Segundos de espera cuando error.retryable=true; no repetir una acción con versión nueva sin decisión.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          }
        }
      },
      "Error422": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Error429": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Segundos de espera cuando error.retryable=true; no repetir una acción con versión nueva sin decisión.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          }
        }
      },
      "Error500": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Error503": {
        "description": "Error normalizado; códigos y condiciones en docs/api/contracts.md.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Segundos de espera cuando error.retryable=true; no repetir una acción con versión nueva sin decisión.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          }
        }
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "manager",
          "operator"
        ]
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "VALIDATION_ERROR",
          "UNAUTHENTICATED",
          "FORBIDDEN",
          "NOT_FOUND",
          "VERSION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "REQUEST_IN_PROGRESS",
          "INVALID_ORDER_TRANSITION",
          "QUOTE_CHANGED",
          "QUOTE_EXPIRED",
          "PRODUCT_UNAVAILABLE",
          "DELIVERY_UNAVAILABLE",
          "BUSINESS_CLOSED",
          "HANDOFF_REQUIRED",
          "WINDOW_CLOSED",
          "RATE_LIMITED",
          "PROVIDER_UNAVAILABLE",
          "INTERNAL_ERROR"
        ]
      },
      "Meta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "request_id"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string",
                "maxLength": 500
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "field": {
                      "type": "string",
                      "maxLength": 120
                    },
                    "issue": {
                      "type": "string",
                      "maxLength": 200
                    }
                  },
                  "required": [
                    "field",
                    "issue"
                  ]
                },
                "maxItems": 20
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "details",
              "retryable"
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "error",
          "meta"
        ]
      },
      "Business": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "business_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "maxLength": 120
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "timezone": {
            "type": "string",
            "maxLength": 120
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ]
          }
        },
        "required": [
          "id",
          "business_id",
          "created_at",
          "updated_at",
          "version",
          "name",
          "slug",
          "currency",
          "timezone",
          "status"
        ]
      },
...
      "Membership": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "business_id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "business_id",
          "role"
        ]
      },
      "Me": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Membership"
            },
            "maxItems": 100
          }
        },
        "required": [
          "user_id",
          "memberships"
        ]
      },
...
      "BusinessUpdate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "timezone": {
            "type": "string",
            "maxLength": 120
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          }
        },
        "required": [
          "expected_version"
        ],
        "minProperties": 2
      },
...
      "BusinessResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Business"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "MeResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Me"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}